+++ InterneTaakAfhandeling.Web.Server/Data/Entities/BaseEntity.cs
@@ -1,0 +1,13 @@
+using System.ComponentModel.DataAnnotations;
+
+namespace InterneTaakAfhandeling.Web.Server.Data.Entities;
+
+public class BaseEntity
+{
+    [Key]
+    public Guid Id { get; set; }
+
+    public DateTimeOffset CreatedAt { get; set; } 
+
+    public DateTimeOffset UpdatedAt { get; set; }
+}